\name{AwsSdkCrt-package}
\alias{AwsSdkCrt-package}
\alias{AwsSdkCrt}
\docType{package}
\title{\packageTitle{AwsSdkCrt}}
\description{\packageDescription{AwsSdkCrt}}
\section{Package Content}{\packageIndices{AwsSdkCrt}}
\author{\packageAuthor{AwsSdkCrt}}
\section{Maintainer}{\packageMaintainer{AwsSdkCrt}}
\keyword{package}
